+++ NEWS.md
@@ -7,6 +7,7 @@
 ## Changes to functions
 
 * `get_dv_labels()` gets a `multi.resp`-argument to return each label of a multivariate response model (only for _brmsfit_ objects).
+* `get_label()` now also returns name-attribute for empty labels if `x` was a data.frame.
 
 ## Bug fixes